Laboratoare/Laborator_2/requirements.txt
requests>=2.28.0
redis>=4.5.0
dicttoxml>=1.7.4
cassandra-driver>=3.25.0